writer: introduce format version 1.2
Introduce "version 2" composefs format by bumping the maximum version.
We intend to use this version to break with some decisions that were
taken for historical reasons and to take advantage of some features that
are only available in newer kernel versions.

There are no changes introduced yet, but the following commits will add
them.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>