writer: set 32bit inode equal to 64bit inode
We write a 'i_ino' field into the erofs inode structs for compatibility
with 32bit systems or 32bit applications running on 64bit systems.  This
is done because the 'nid' that erofs uses for the inode ID on 64bit
might overflow a 32bit integer.

In order for that to happen, though, the inode table alone would have to
grow past 128GiB in size: since inode alignment is 32 bytes, the nid
*with a maximum value of 4Gi) is multiplied by 32B to determine the file
offset of the inode.

This seems exceedingly unlikely to ever occur.  composefs images for
even very large systems are on the order of less than 100MB of total
size, with the inode table even smaller than that.

Let's just set the 32bit inode value the same as the 64bit value.

This also avoids another potential issue: under the old approach, we use
0 as the 32bit value for the root directory inode (which always comes
first).  Some parts of the kernel (`grep is_zero_ino`, for example)
claim that userspace might depend on inodes being non-zero, and all
other filesystems seem to respect this.  nids are always non-zero (since
an inode at the start of the file would overlap the header).

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>