writer: don't store xattr block in superblock
In erofs, finding a shared xattr involves taking the 32bit id,
multiplying it by 4, and using that as an index into the section of the
file starting at the block named for shared xattr data in the
superblock (superblock.xattr_blkaddr).

This is only useful of the inode and xattr data stored in the filesystem
would exceed 16GiB, which is exceedingly unlikely to ever occur for a
composefs.

We can make our lives a fair bit easier if we set the xattr starting
block to 0 and simply store the file offset (divded by 4) as the xattr
id directly.

There is a similar pointer (superblock.meta_blkaddr) for the start of
the inodes.  This calculation only involves constants and always works
out to zero.  We could delete that entirely, but in any case, let's not
bother with it for version 2.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>